split CORSFilter file and add scaladoc
